Bugfix in function, re-do some dependencies
vim-autosurround is kinda slow and doesn't work as expected in several cases, as well as being buggy if Python 2 is missing (python 2 alone is a bad sign, considering it's at the end of its lifespan). Added codi to the list - might be useful on Linux, or if it one day starts supporting Windows. The python dependencies now also expect a binary somewhere with python2/python3 as its name. This is to generify the executable location (because it still works, even if it doesn't use a full path). Neovim's :CheckHealth triggers on that though, but Neovim isn't my primary use.