Adds zsh.
I've really come to like zsh. Messed around with it a little on my
older computer, and it's really good. It's fast, looks pretty, and adds
a whole bunch of stuff that makes the command line really good. That
matters because I spend a lot of time on the command line. This is a
bootstrap change so I can install the basics on my other computer, and
figure out anything missing in this change.