Fix more stuff MS should've fixed over half a decade ago, beef up dotfiles for the work machine reload
Closes #16
Initial work done in 84f93f9136242579bffcbca426950e04bfe269a5, per
    https://github.com/microsoft/wslg/issues/173,
    extended this time with automation.

The WSL keyboard hack can indeed be an in-WSL bash script, and it's now
sourced normally along with the zshrc. This eliminates the manual
per-boot work I needed to do for WSL GUI apps to be usable with a dead
keys layout.