Drop startify
Startify's code is unsalvageable. The amount of rewriting required to
make it even vaguely decent for development greatly exceeds the effort
it would take to write a new plugin from scratch. I don't think it makes
sense to write yet another plugin for it, so I think I'm just going to
roll with the default start screen.

In the long term, it might be worth it to do so, but we are not there
yet. I first need an idea of what I'd want in a Vim dashboard or start
page to actually justify investing potentially huge amounts of time in a
thing I currently only see for a couple seconds at most.

Related: #25