Switch away from clang by default
I did not think this day would ever come, but here it is.

I've been experimenting more with coverage, and clang simply does not
have the same trivial support. Most tutorials and similar use
gcov/lcov/genhtml, which is fairly standard tooling, but doesn't work
with clang without a .sh wrapper that's also included in this commit.
Unfortunately, I would have to symlink it to `~/.local/bin/gcov` for it
to work with standard `find-package(... gcov)` statements, which isn't
great.

Therefore, it's time to switch back to GCC. It simply has better tooling
support and requires less fuckery for the tooling to work. I'm not even
sure I'll be staying in the ecosystem in general for much longer.
Clang-tidy will likely stick, but clangd might get thrown out in favour
of ccls. I'll be experimenting with it Soon:tm:, so that's still
pending.